Move logic of fetching elm.json files to a separate module
