+++ lib/project-dependencies.js
@@ -3,6 +3,7 @@ const util = require('util');
 const os = require('os');
 const got = require('got');
 const fs = require('fs-extra');
+const ProjectJsonFiles = require('./project-json-files');
 
 const fsMkdirp = util.promisify(fs.mkdirp);
 const fsReadJson = util.promisify(fs.readJson);
@@ -60,7 +61,7 @@ async function collect(options, elmJson, reviewDirectDependencies, elmVersion) {
           hasDependenciesThatCouldNotBeDownloaded = true;
           return [];
         }),
-        getElmJson(
+        ProjectJsonFiles.getElmJson(
           elmVersion,
           elmReviewDependencyCache(name, packageVersion),
           name,
@@ -79,7 +80,9 @@ async function collect(options, elmJson, reviewDirectDependencies, elmVersion) {
     }
   );
 
-  const projectDeps = await Promise.all(projectDepsPromises).then((items) => items.filter(Boolean));
+  const projectDeps = await Promise.all(projectDepsPromises).then((items) =>
+    items.filter(Boolean)
+  );
 
   if (hasDependenciesThatCouldNotBeDownloaded && options.report !== 'json') {
     console.log(
@@ -101,7 +104,11 @@ async function collectRuleLinks(reviewDirectDependencies, elmVersion) {
   )
     .filter(([name]) => !name.startsWith('elm/'))
     .map(([name, packageVersion]) =>
-      getElmJsonFromElmHome(elmVersion, name, packageVersion).catch(() => null)
+      ProjectJsonFiles.getElmJsonFromElmHome(
+        elmVersion,
+        name,
+        packageVersion
+      ).catch(() => null)
     );
 
   const elmJsonForReviewDependencies = await Promise.all(
@@ -168,54 +175,6 @@ function getDocsJson(
     );
 }
 
-function getElmJson(
-  elmVersion,
-  elmReviewDependencyCache,
-  name,
-  packageVersion
-) {
-  // Look for the dependency in ELM_HOME first
-  return (
-    getElmJsonFromElmHome(elmVersion, name, packageVersion)
-      // Then in the dependency cache for elm-review
-      .catch(() => fsReadJson(path.join(elmReviewDependencyCache, 'elm.json')))
-      .catch(() =>
-        // Finally, try to download it from the packages website
-        readElmJsonFromPackagesWebsite(name, packageVersion).then((result) => {
-          // And cache it for future uses
-          cacheFile(
-            elmReviewDependencyCache,
-            'elm.json',
-            result
-          ).catch(() => {});
-          return result;
-        })
-      )
-  );
-}
-
-// TODO Empty this cache at some point?
-// Note that we might need it in watch and fix mode, but not otherwise.
-const elmJsonInElmHomePromises = new Map();
-function getElmJsonFromElmHome(elmVersion, name, packageVersion) {
-  const key = `${elmVersion}-${name}-${packageVersion}`;
-  let promise = elmJsonInElmHomePromises.get(key);
-  if (promise) {
-    return promise;
-  }
-
-  const directory = path.join(
-    elmRoot,
-    elmVersion,
-    'packages',
-    name,
-    packageVersion
-  );
-  promise = fsReadJson(path.join(directory, 'elm.json'));
-  elmJsonInElmHomePromises.set(key, promise);
-  return promise;
-}
-
 async function cacheFile(directory, name, jsonContent) {
   await fsMkdirp(directory).catch(() => {});
   return fsWriteJson(path.join(directory, name), jsonContent);
@@ -241,13 +200,6 @@ function defaultElmJson(name, packageVersion) {
 async function readDocsJsonFromPackagesWebsite(packageName, packageVersion) {
   const response = await got(
     `https://package.elm-lang.org/packages/${packageName}/${packageVersion}/docs.json`
-  );
-  return JSON.parse(response.body);
-}
-
-async function readElmJsonFromPackagesWebsite(packageName, packageVersion) {
-  const response = await got(
-    `https://package.elm-lang.org/packages/${packageName}/${packageVersion}/elm.json`
   );
   return JSON.parse(response.body);
 }
