+++ lib/runner.js
@@ -142,7 +142,7 @@ async function initializeApp(options, elmModulePath, reviewElmJson, appHash) {
     projectDepsP,
     suppressedErrorsP,
     linksToRuleDocsP
-  ])
+  ]);
 
   await sendProjectContent(
     options,
